bowman_py_server/requirements.txt
requests==2.25.1
uvicorn==0.13.4
pydantic==1.8.1
fastapi==0.63.0
gevent_socketio==0.3.6
python-dotenv==0.15.0
